Change struct CFX_BidiChar::Segment to have size_t fields
The `start` and `count` fields index into containers and are never
negative. Switch them to be size_t so they can be more easily used with
spans. e.g. in https://pdfium-review.googlesource.com/c/151032

TAG=agy
CONV=117163e4-c12b-43d7-9709-b718d4450d3e

Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/151032
Commit-Queue: Lei Zhang <thestig@chromium.org>

+++ core/fpdftext/cpdf_textpage.cpp
@@ -847,14 +847,14 @@ void CPDF_TextPage::CloseTempLine() {
         (segment.direction == CFX_BidiChar::Direction::kNeutral &&
          current_direction == CFX_BidiChar::Direction::kRight)) {
       current_direction = CFX_BidiChar::Direction::kRight;
-      for (int m = segment.start + segment.count; m > segment.start; --m) {
+      for (size_t m = segment.start + segment.count; m > segment.start; --m) {
         AddCharInfo(str[m - 1], temp_char_list_[m - 1], /*is_rtl=*/true);
       }
     } else {
       if (segment.direction != CFX_BidiChar::Direction::kLeftWeak) {
         current_direction = CFX_BidiChar::Direction::kLeft;
       }
-      for (int m = segment.start; m < segment.start + segment.count; ++m) {
+      for (size_t m = segment.start; m < segment.start + segment.count; ++m) {
         AddCharInfo(str[m], temp_char_list_[m], /*is_rtl=*/false);
       }
     }
